backend_model/requirements.txt
flask
torch
transformers
pillow
gunicorn
flask-cors